--- src/ts/jsexec/index.ts
@@ -1,193 +1,0 @@
-/**
- * ArcOS JavaScript Execution Engine
- *
- * This file executes JS files in ArcOS under a relatively controlled environment.
- * It is part of the ArcOS TPA framework: a system for running third-party apps.
- *
- * All rights belong to their respective authors.
- *
- * © IzKuipers 2025
- */
-import type { ITpaConnector } from "$interfaces/modules/server/ITpaConnector";
-import { ThirdPartyAppProcess } from "$ts/apps/thirdparty";
-import { ThirdPartyProps } from "$ts/apps/tpa/props";
-import { Daemon, Env, Fs, Stack } from "$ts/env";
-import { Process } from "$ts/kernel/mods/stack/process/instance";
-import { arrayBufferToText } from "$ts/util/convert";
-import { getItemNameFromPath, getParentDirectory } from "$ts/util/fs";
-import { UUID } from "$ts/util/uuid";
-import type { App } from "$types/apps/app";
-import type { ParsedImportStatement, ThirdPartyPropMap } from "$types/tpa/thirdparty";
-import * as acorn from "acorn";
-
-export class JsExec extends Process {
-  public readonly TPA_REVISION = ThirdPartyAppProcess.TPA_REV;
-  props?: ThirdPartyPropMap;
-  app?: App;
-  args: any[];
-  metaPath?: string;
-  filePath?: string;
-  workingDirectory: string;
-  operationId: string;
-
-  //#region LIFECYCLE
-
-  constructor(pid: number, parentPid: number, filePath: string, ...args: any[]) {
-    super(pid, parentPid);
-
-    this.args = args;
-    this.filePath = filePath;
-    this.workingDirectory = getParentDirectory(filePath);
-    this.name = "JsExec";
-    this.setSource(__SOURCE__);
-    this.operationId = UUID();
-  }
-
-  async start() {
-    if (!this.filePath) return false;
-
-    this.props = ThirdPartyProps(this);
-  }
-
-  //#endregion
-  //#region URL
-
-  private async getTpaUrl(wrapped: string) {
-    this.Log(`Getting TPA file URL`);
-
-    const { appId, userId, filename } = this.getTpaUrlInfo();
-    try {
-      const urlResult = await Daemon!.GetConnector<ITpaConnector>("TpaConnector").CreateUrl(wrapped, userId, appId, filename);
-
-      if (!urlResult.success) throw new JsExecError();
-      return urlResult.result!;
-    } catch (e: any) {
-      throw new JsExecError(`Failed to create momentary TPA URL: ${e?.message ?? e}`);
-    }
-  }
-
-  private getTpaPostUrl() {
-    const { appId, userId, filename } = this.getTpaUrlInfo();
-
-    return `/tpa/v2/${userId}/${appId}/${filename}`;
-  }
-
-  private getTpaUrlInfo() {
-    const appId = this.app?.id || "ArcOS";
-    const userId = Daemon?.userInfo?._id || "SYSTEM";
-    const filename = getItemNameFromPath(this.filePath!);
-
-    return { appId, userId, filename };
-  }
-
-  //#endregion
-  //#region EXECUTION
-
-  private async exec(tpaUrl: string) {
-    this.Log(`Executing ${this.filePath}`);
-
-    const code = await import(/* @vite-ignore */ tpaUrl);
-
-    if (!code.default || !(code.default instanceof Function)) throw new JsExecError("Expected a default function");
-
-    try {
-      const result = await code.default(this.props!);
-      return result;
-    } catch (e) {
-      throw e;
-    } finally {
-      await this.killSelf();
-    }
-  }
-
-  async getContents() {
-    this.Log(`Reading script contents`);
-
-    const unwrapped = this.convertImportStatementsToRegex(arrayBufferToText((await Fs.readFile(this.filePath!))!)!);
-    if (!unwrapped) throw new JsExecError(`Failed to read ${this.filePath}: not found`);
-
-    await this.testFileContents(unwrapped);
-
-    const wrapped = this.wrap(unwrapped);
-    const tpaUrl = await this.getTpaUrl(wrapped);
-
-    return await this.exec(tpaUrl);
-  }
-
-  //#endregion
-  //#region HELPERS
-
-  setApp(app: App, metaPath?: string) {
-    this.Log(`Setting app data to ${app.id} (${metaPath ?? "<unknown meta>"})`);
-
-    if (this.app) return;
-
-    if (app.tpaRevision && app.tpaRevision > this.TPA_REVISION)
-      throw new JsExecError(
-        `This application expects a newer version of the TPA framework than what ArcOS can supply. Please update your ArcOS version and try again.`
-      );
-
-    this.app = app;
-    this.metaPath = metaPath;
-    this.props = ThirdPartyProps(this);
-  }
-
-  private wrap(contents: string) {
-    if (!this.props) throw new JsExecError(`No TPA props to use`);
-
-    return `export default async function({${Object.keys(this.props).join(",")}}) {\nconst global = arguments;\n${contents}\n}`;
-  }
-
-  private convertImportStatementsToRegex(sourceFile: string) {
-    if (!sourceFile) return sourceFile;
-    const regex =
-      /import(?:(?:(?:[ \n\t]+(?<default>[^ *\n\t\{\},]+)[ \n\t]*(?:,|[ \n\t]+))?(?<destructured>[ \n\t]*\{(?:[ \n\t]*[^ \n\t"'\{\}]+[ \n\t]*,?)+\})?[ \n\t]*)|[ \n\t]*\*[ \n\t]*as[ \n\t]+(?<wildcard>[^ \n\t\{\}]+)[ \n\t]+)from[ \n\t]*(?:['"])(?<filename>[^'"\n]+)(?<quote>['"])/gm;
-    const matches = sourceFile
-      .matchAll(regex)
-      .toArray()
-      .map((m) => ({ ...m.groups, original: m[0] })) as ParsedImportStatement[];
-
-    for (const match of matches) {
-      const { destructured, default: defaultImport, filename, quote, wildcard } = match;
-      let loadStatement = `const ${destructured || defaultImport || wildcard} = await load(${quote}${filename}${quote})`;
-
-      sourceFile = sourceFile.replace(match.original, loadStatement);
-    }
-
-    return sourceFile;
-  }
-
-  private async testFileContents(unwrapped: string) {
-    try {
-      const ast = acorn.parse(unwrapped, {
-        sourceType: "module",
-        ecmaVersion: "latest",
-        allowReturnOutsideFunction: true,
-        allowAwaitOutsideFunction: true,
-      });
-      const hasExport = ast.body.some((node) => node.type.startsWith("Export"));
-      const hasImport = ast.body.some((node) => node.type.startsWith("Import"));
-      const hasDebugger = ast.body.some((node) => node.type.startsWith("Debugger"));
-
-      if (hasExport) throw new JsExecError("Export statements are not valid inside of ArcOS");
-      if (hasImport) throw new JsExecError("Import statements are not valid inside of ArcOS");
-      if (hasDebugger) throw new JsExecError("Debugger triggers are not valid inside of ArcOS");
-    } catch (e) {
-      throw new JsExecError(`An error occurred while parsing the source file: ${e}`);
-    }
-  }
-
-  //#endregion
-
-  static async Invoke(filePath: string, ...args: any[]) {
-    return await Stack.spawn<JsExec>(JsExec, undefined, undefined, +Env.get("userdaemon_pid"), filePath, ...args);
-  }
-}
-
-export class JsExecError extends Error {
-  constructor(message?: string, options?: ErrorOptions) {
-    super(message, options);
-
-    this.name = "JsExecError";
-  }
-}
