--- src/ts/servicehost/services/JsExec/props.ts
@@ -1,134 +1,0 @@
-import type { IThirdPartyAppProcess } from "$interfaces/IThirdPartyAppProcess";
-import { AppProcess } from "$ts/apps/process";
-import { ThirdPartyAppProcess } from "$ts/apps/thirdparty";
-import { __Console__ } from "$ts/console";
-import { Daemon, Env, Fs, Stack } from "$ts/env";
-import { getAllImages } from "$ts/images";
-import { FilesystemDrive } from "$ts/kernel/mods/fs/drives/generic";
-import { Backend } from "$ts/kernel/mods/server/axios";
-import { Process } from "$ts/kernel/mods/stack/process/instance";
-import { BaseService } from "$ts/servicehost/base";
-import { Sleep } from "$ts/sleep";
-import { contextProps } from "$ts/ui/context/actions.svelte";
-import { CustomTitlebar } from "$ts/ui/thirdparty/titlebar";
-import { TrayIconProcess } from "$ts/ui/tray/process";
-import { HiddenUserPaths, SystemFolders, UserPathCaptions, UserPathIcons, UserPaths } from "$ts/user/store";
-import { CountInstances, decimalToHex, htmlspecialchars, Plural, sha256, sliceIntoChunks } from "$ts/util";
-import { arrayBufferToBlob, arrayBufferToText, blobToDataURL, blobToText, textToArrayBuffer, textToBlob } from "$ts/util/convert";
-import { BTN_OKAY_SUG, MessageBox } from "$ts/util/dialog";
-import {
-  DownloadFile,
-  formatBytes,
-  getDriveLetter,
-  getItemNameFromPath,
-  getParentDirectory,
-  join,
-  onFileChange,
-  onFolderChange,
-} from "$ts/util/fs";
-import { tryJsonStringify } from "$ts/util/json";
-import { Store } from "$ts/writable";
-import type { ThirdPartyPropMap } from "$types/tpa/thirdparty";
-import axios from "axios";
-import dayjs from "dayjs";
-import { ThirdPartyProcess } from "$ts/apps/tpa/process";
-import { SupplementaryThirdPartyPropFunctions } from "./supplementary";
-import type { JsExecEngineData } from "./engine";
-
-export function ThirdPartyProps(engine: JsExecEngineData): ThirdPartyPropMap {
-  const props = {
-    env: Env, // TEMP
-    handler: Stack, // TEMP
-    fs: Fs, // TEMP
-    daemon: Daemon, // TEMP
-    serviceHost: Daemon!.serviceHost, // TEMP
-    MessageBox,
-    icons: getAllImages(),
-    util: {
-      htmlspecialchars,
-      Plural,
-      sliceIntoChunks,
-      decimalToHex,
-      sha256,
-      CountInstances,
-      join,
-      getItemNameFromPath,
-      getParentDirectory,
-      getDriveLetter,
-      formatBytes,
-      DownloadFile,
-      onFileChange,
-      onFolderChange,
-    },
-    convert: {
-      arrayToText: arrayBufferToText,
-      textToArrayBuffer,
-      blobToText,
-      textToBlob,
-      arrayToBlob: arrayBufferToBlob,
-      blobToDataURL,
-    },
-    workingDirectory: engine.workingDirectory || engine.app?.workingDirectory!,
-    Process,
-    AppProcess,
-    ThirdPartyAppProcess,
-    ThirdPartyProcess,
-    FilesystemDrive,
-    argv: engine.args,
-    app: engine.app,
-    Store,
-    Sleep,
-    $ENTRYPOINT: engine.filePath,
-    $METADATA: engine.metaPath,
-    SHELL_PID: +Env.get("shell_pid"),
-    OPERATION_ID: engine.operationId,
-    load: async (path: string): Promise<any> => {},
-    runApp: async (
-      process: typeof ThirdPartyAppProcess,
-      metadataPath: string,
-      parentPid?: number,
-      ...args: any[]
-    ): Promise<IThirdPartyAppProcess | undefined> => undefined,
-    loadHtml: async (path: string): Promise<string | undefined> => undefined,
-    axios,
-    Server: Backend,
-    BaseService,
-    TrayIconProcess,
-    Debug: (m: any) => {
-      MessageBox(
-        {
-          title: "🐛🪵",
-          message: tryJsonStringify(m, 2),
-          image: "WindowSettingsIcon",
-          sound: "arcos.dialog.info",
-          buttons: [BTN_OKAY_SUG],
-        },
-        +Env.get("shell_pid")
-      );
-    },
-    CustomTitlebar,
-    contextProps,
-    dayjs,
-    console: __Console__,
-    LogLevel: {
-      info: 0,
-      warning: 1,
-      error: 2,
-      critical: 3,
-    },
-    UserPaths,
-    UserPathCaptions,
-    UserPathIcons,
-    SystemFolders,
-    HiddenUserPaths,
-  };
-
-  const supplementary = SupplementaryThirdPartyPropFunctions(engine);
-
-  for (const [key, supp] of Object.entries(supplementary)) {
-    (props as any)[key] = supp;
-  }
-
-  //@ts-ignore
-  return props;
-}
